Convert SurgeonGeneralTafani plugin from kts to kt
